Add OneSplineOrbData.hpp
